Remove WC Support
Removed WC support as it seems to conflict with current WC versions and requires updates. Moreover, the theme should be minimal, in future more will be removed (such as BS library) since Toolset already provides that.
As well in this change, a duplicate Filter was removed (integrating Layouts) which was called twice - my bad, who's else 🗡 
Props to https://toolset.com/forums/topic/issues-when-importing-layouts-exports/ for discovering it.